Adds `--plays` option to add play count in front of file names when downloading from SoundCloud playlists.
Why? Because sometimes when you want to archive a very long play list (eg https://soundcloud.com/muloka/sets/burning-man-2016) it could be handy to be able to sort files by "popularity". To achieve that, `--plays` option adds play count at the moment of the download as the first thing in the file name.

If you want to resync the playlist, files that were downloaded already won't be redownloaded with new play count. At the same time, existing file's play count won't be updated either.